revised shipment (& commented out course controller)

+++ src/main/java/com/example/cms/model/repository/ShipmentRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ShipmentRepository extends JpaRepository<Shipment, Long> {
+public interface ShipmentRepository extends JpaRepository<Shipment, Integer> {
 
     @Query(value = "select * from Shipment S", nativeQuery = true)
     List<Shipment> search(@Param("shipmentId") int shipmentId);
